pyproject.toml
[project]
name = "quantumstate"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "certstream>=1.12",
    "cryptography>=46.0.5",
    "datasets>=4.5.0",
    "elasticsearch>=8.11.0",
    "fastapi>=0.111.0",
    "ipykernel>=7.2.0",
    "jupyter>=1.1.1",
    "pandas>=2.0.0,<3",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "sse-starlette>=1.8.2",
    "streamlit>=1.40.0",
    "uvicorn[standard]>=0.29.0",
]
